requirements.txt
flask
torch
torchvision
opencv-python-headless
albumentations
pydicom
pandas
numpy
grad-cam
pillow

pip install -r requirements.txt